Move ListItem post creation to extension
