+++ Example/SpotsFeed/SpotsFeed/Cells/CommentTableViewCell.swift
@@ -10,7 +10,7 @@ public class CommentTableViewCell: WallTableViewCell, Itemble {
   public var size = CGSize(width: 0, height: 44)
 
   public class func height(item: ListItem) -> CGFloat {
-    let post = CommentTableViewCell.itemToPost(item)
+    let post = item.post
     let postText = post.text as NSString
     let textFrame = postText.boundingRectWithSize(CGSize(
       width: UIScreen.mainScreen().bounds.width - Dimensions.textOffset - Dimensions.sideOffset,
@@ -123,21 +123,6 @@ public class CommentTableViewCell: WallTableViewCell, Itemble {
   public required init?(coder aDecoder: NSCoder) {
     fatalError("init(coder:) has not been implemented")
   }
-  
-  private class func itemToPost(item: ListItem) -> Post {
-    let avatarURL = NSURL(string: item.image)!
-    let author = Author(name: item.title, avatar: avatarURL)
-
-    var mediaItems = [Media]()
-    if let strings = item.meta["media"] as? [String] {
-      for mediaString in strings {
-        let url = NSURL(string: mediaString)!
-        let media = Media(kind: Media.Kind.Image, source: url)
-        mediaItems.append(media)
-      }
-    }
-    return Post(id: 0, text: item.subtitle, publishDate: "", author: author, media: mediaItems)
-  }
 
   // MARK: - Actions
 
@@ -149,9 +134,7 @@ public class CommentTableViewCell: WallTableViewCell, Itemble {
   // MARK: - Setup
 
   public func setupViews(item: ListItem) -> CGFloat {
-    post = CommentTableViewCell.itemToPost(item)
-    guard let post = post else { return 0 }
-
+    let post = item.post
     let totalWidth = UIScreen.mainScreen().bounds.width
 
     avatarImageView.frame = CGRect(x: Dimensions.sideOffset, y: Dimensions.sideOffset,
@@ -184,8 +167,7 @@ public class CommentTableViewCell: WallTableViewCell, Itemble {
       item.size.width = contentView.frame.width
       item.size.height = setupViews(item)
     } else {
-      post = CommentTableViewCell.itemToPost(item)
-      item.size.height = CommentTableViewCell.height(item)
+      item.size.height = CommentTableViewCell.height(item.post)
     }
   }
 }
